major upgrade firebase package, init firebase app at one place only
