+++ package.json
@@ -24,7 +24,7 @@
     "expo-font": "~11.1.1",
     "expo-status-bar": "~1.4.4",
     "expo-updates": "~0.16.4",
-    "firebase": "^9.20.0",
+    "firebase": "^10.7.1",
     "react": "18.2.0",
     "react-hook-form": "^7.50.1",
     "react-native": "0.71.6",
