--- src/firebase/firebaseConfig.js
@@ -1,12 +1,0 @@
-
-
-export default firebaseConfig = {
-  apiKey: "AIzaSyBS1fCY-_I-V4WZYveOTNmx9ByGDRaCW70",
-  authDomain: "pre-phrase.firebaseapp.com",
-  databaseURL: "https://pre-phrase-default-rtdb.asia-southeast1.firebasedatabase.app",
-  projectId: "pre-phrase",
-  storageBucket: "pre-phrase.appspot.com",
-  messagingSenderId: "531222180809",
-  appId: "1:531222180809:web:91c3af0b3779e82801ee18",
-  measurementId: "G-1EJ9733JN2"
-};
